services/transcription-service/requirements.txt
websockets==12.0
paho-mqtt==2.1.0
pydub==0.25.1